add to EditSupplierFragment.kt update supplier view

+++ app/src/main/java/com/example/salestapapp/supplier/data/database/dao/SuppliersDao.kt
@@ -23,6 +23,6 @@ interface SuppliersDao {
     suspend fun deleteSupplierByID(supplierID: Int)
 
     @Update
-    suspend fun editSupplier(suppliers: SupplierEntity):Int
+    suspend fun editSupplier(suppliers: SupplierEntity):Long
 
 }
